fix : vercel 배포를 위한 수정
