+++ src/app/advertiser/ads-list/register/styles/FormSections/ParticipationInfoSection.module.css
@@ -1,6 +1,6 @@
 .formSection {
   background: white;
-  border: 1px solid #E5E7EB;
+  border: 1px solid #e5e7eb;
   border-radius: 12px;
   padding: 25px;
   box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.05);
@@ -10,12 +10,12 @@
   display: block;
   font-size: 18px;
   font-weight: 400;
-  color: #594A3E;
+  color: #594a3e;
   margin-bottom: 16px;
 }
 
 .required {
-  color: #EF4444;
+  color: #ef4444;
   margin-left: 4px;
 }
 
@@ -35,7 +35,7 @@
 .dateLabel {
   font-size: 15px;
   font-weight: 400;
-  color: #594A3E;
+  color: #594a3e;
 }
 
 .dateInputs {
@@ -45,51 +45,56 @@
 }
 
 .dateInputs span {
-  color: #6B7280;
+  color: #6b7280;
   font-size: 14px;
 }
 
 .datePicker {
   flex: 1;
   padding: 16px;
-  border: 1px solid #D1D5DB;
+  border: 1px solid #d1d5db;
   border-radius: 8px;
   font-size: 16px;
   font-weight: 400;
-  color: #594A3E;
+  color: #594a3e;
   background: white;
   cursor: pointer;
 }
 
 .datePicker:focus {
   outline: none;
-  border-color: #F5A623;
+  border-color: #f5a623;
+}
+
+.datePickerContainer {
+  /* DatePicker 컨테이너 스타일 */
 }
 
 /* DatePicker 커스텀 스타일 */
-:global(.react-datepicker-wrapper) {
+.datePickerContainer :global(.react-datepicker-wrapper) {
   flex: 1;
 }
 
-:global(.react-datepicker__input-container input) {
+.datePickerContainer :global(.react-datepicker__input-container input) {
   width: 100%;
   padding: 16px;
-  border: 1px solid #D1D5DB;
+  border: 1px solid #d1d5db;
   border-radius: 8px;
   font-size: 16px;
   font-weight: 400;
-  color: #594A3E;
+  color: #594a3e;
   background: white;
   cursor: pointer;
 }
 
-:global(.react-datepicker__input-container input:focus) {
+.datePickerContainer :global(.react-datepicker__input-container input:focus) {
   outline: none;
-  border-color: #F5A623;
+  border-color: #f5a623;
 }
 
-:global(.react-datepicker__input-container input::placeholder) {
-  color: #ADAEBC;
+.datePickerContainer
+  :global(.react-datepicker__input-container input::placeholder) {
+  color: #adaebc;
 }
 
 @media (max-width: 768px) {
@@ -101,20 +106,20 @@
 .input {
   width: 100%;
   padding: 16px;
-  border: 1px solid #D1D5DB;
+  border: 1px solid #d1d5db;
   border-radius: 8px;
   font-size: 16px;
   font-weight: 400;
-  color: #594A3E;
+  color: #594a3e;
   background: white;
   box-sizing: border-box;
 }
 
 .input:focus {
   outline: none;
-  border-color: #F5A623;
+  border-color: #f5a623;
 }
 
 .input::placeholder {
-  color: #ADAEBC;
+  color: #adaebc;
 }
